Update firmware.ino
This is a fix for converters such as the SNES2GC so that inputs are accurately displayed. 

This retains the functionality of using the input display on a standard SNES console while also fixing issues with adapters for other consoles.

This wasn't tested on a Raphnet adapter (I do not own one) but this works on two input displays I own through the SNES2GC adapter.